Moved sample experiment to template experiment file
